Add ability to switch between multiple ROM sets
Introduce a ROM manager that scans a 'roms' directory on startup for all
valid ROMs that mt32emu supports.

If none are found in this directory, fall back on searching for
MT32_{CONTROL|PCM}.ROM in the root of the SD card as before.

Add a custom SysEx command that allows the user to switch ROM sets on
the fly.

Add a new config file option for setting the default ROM set.

Closes #18.